fix comments

+++ openrtb_ext/imp_intertech.go
@@ -1,6 +1,6 @@
 package openrtb_ext
 
 type ExtImpIntertech struct {
-	PageID int64 `json:"page_id"`
-	ImpID  int64 `json:"imp_id"`
+	PageID int `json:"page_id"`
+	ImpID  int `json:"imp_id"`
 }
